Gear Up: Integrated latest from dev.
[git-p4: depot-paths = "//depot/traktor/": change = 13800]

+++ code/Input/OsX/InputDeviceJoystickOsX.cpp
@@ -186,7 +186,7 @@ void InputDeviceJoystickOsX::callbackValue(void* context, IOReturn result, void*
 		return;
 		
 	uint32_t usage = (uint32_t)IOHIDElementGetUsage(element);
-	if (usage == ~0UL)
+	if (usage == uint32_t(~0UL))
 		return;
 		
 	int32_t v = (int32_t)IOHIDValueGetIntegerValue(value);
